refactor: remove unnecessary nesting
